requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
filelock==3.9.0
huggingface-hub==0.12.1
idna==3.4
numpy==1.24.2
packaging==23.0
psycopg2-binary==2.9.5
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sentencepiece==0.1.97
soupsieve==2.4
sqlparse==0.4.3
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
human-centered-summarization/financial-summarization-pegasus

typing_extensions==4.5.0
urllib3==1.26.14
